+++ packages/desktop/package.json
@@ -3,28 +3,14 @@
   "version": "1.0.0-alpha.2",
   "description": "Visual app using three.js",
   "main": "./out/main/index.js",
-  "types": "./dist/types/index.d.ts",
-  "files": [
-    "dist/types/**/*"
-  ],
-  "exports": {
-    ".": {
-      "types": "./dist/types/index.d.ts",
-      "default": "./out/main/index.js"
-    },
-    "./types": {
-      "types": "./dist/types/index.d.ts"
-    }
-  },
   "scripts": {
     "lint": "eslint . --ext .js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
     "typecheck:node": "tsc --noEmit -p tsconfig.node.json --composite false",
     "typecheck:web": "tsc --noEmit -p tsconfig.web.json --composite false",
     "typecheck": "pnpm run typecheck:node && pnpm run typecheck:web",
-    "build:types": "tsc --project tsconfig.types.json",
     "start": "electron-vite preview",
     "dev": "electron-vite dev",
-    "build": "pnpm run build:types && electron-vite build",
+    "build": "electron-vite build",
     "build:unpack": "pnpm run build && electron-builder --dir",
     "build:win": "pnpm run build && electron-builder --win",
     "build:mac": "electron-vite build && electron-builder --mac",
@@ -39,12 +25,12 @@
     "@esbuild/linux-x64": "^0.23.1",
     "@esbuild/win32-ia32": "^0.23.1",
     "@esbuild/win32-x64": "^0.23.1",
+    "@hedron/app-store": "workspace:^",
     "@hedron/clock": "workspace:^",
     "@hedron/engine": "workspace:^",
     "@hedron/lfo-input": "workspace:^",
     "@hedron/midi-input": "workspace:^",
     "@hedron/ui-core": "workspace:^",
-    "@redux-devtools/extension": "^3.3.0",
     "@tomjs/electron-devtools-installer": "^2.4.0",
     "chokidar": "^4.0.3",
     "date-fns": "^4.1.0",
@@ -52,7 +38,6 @@
     "esbuild": "^0.23.1",
     "fs-extra": "^11.3.1",
     "get-port-please": "^3.2.0",
-    "immer": "^10.1.1",
     "lodash.debounce": "^4.0.8",
     "path-browserify": "^1.0.1",
     "postprocessing": "^6.37.7",
