--- packages/desktop/src/index.ts
@@ -1,10 +1,0 @@
-// Export the store and its types
-export { appStore, useAppStore } from './renderer/appStore'
-export type {
-  AppState,
-  AppStore,
-  SetState,
-  CustomSetState,
-  DialogId,
-  SaveItem,
-} from './renderer/appStore'
