+++ packages/desktop/src/main/handlers/saveProjectFile.ts
@@ -1,8 +1,8 @@
 import fs from 'fs'
 import path from 'path'
 import { dialog } from 'electron'
+import { ProjectData } from '@hedron/app-store'
 import { SaveProjectResponse } from '@shared/Events'
-import { ProjectData } from '@shared/types'
 
 const isSubdirectory = (parentDir: string, directory: string) => {
   const resolvedParentDir = path.resolve(parentDir)
