+++ packages/desktop/src/main/index.ts
@@ -1,8 +1,8 @@
 import { app, BrowserWindow, dialog, ipcMain, screen, session } from 'electron'
 import { electronApp, optimizer } from '@electron-toolkit/utils'
 import { REDUX_DEVTOOLS, installExtension } from '@tomjs/electron-devtools-installer'
+import { ProjectData } from '@hedron/app-store'
 import { saveFrameHandler, saveFrameSequenceHandler } from './handlers/frameHandlers'
-import { ProjectData } from '@shared/types'
 import {
   DialogEvents,
   FileEvents,
