+++ packages/desktop/src/renderer/components/hooks/useSelectedParam.ts
@@ -1,7 +1,7 @@
 import { Param } from '@hedron/engine'
+import { useEngineStore } from '@hedron/ui-core'
 import { useAppStore } from '@renderer/appStore'
 import { useActiveSketch } from '@components/hooks/useActiveSketch'
-import { useEngineStore } from '@renderer/engine'
 
 export const useSelectedParam = (): Param | null => {
   const activeSketch = useActiveSketch()
