+++ packages/desktop/src/renderer/components/hooks/useSketchList.ts
@@ -1,5 +1,5 @@
+import { useEngineStore } from '@hedron/ui-core'
 import { useMemo } from 'react'
-import { useEngineStore } from '@renderer/engine'
 
 export const useSketchList = () => {
   const sketches = useEngineStore((state) => state.sketches)
