+++ packages/desktop/src/renderer/components/hooks/useSketchModuleList.ts
@@ -1,5 +1,5 @@
+import { useEngineStore } from '@hedron/ui-core'
 import { useMemo } from 'react'
-import { useEngineStore } from '@renderer/engine'
 
 export const useSketchModuleList = () => {
   const modules = useEngineStore((state) => state.sketchModules)
