--- packages/desktop/src/renderer/components/hooks/useUpdateNodeValue.ts
@@ -1,13 +1,0 @@
-import { useEngineStore } from '@renderer/engine'
-
-export const useUpdateNodeValue = () => {
-  const updateNodeValue = useEngineStore((state) => state.updateNodeValue)
-
-  return updateNodeValue
-}
-
-export const useUpdateMultipleNodeValues = () => {
-  const updateMultipleNodeValues = useEngineStore((state) => state.updateMultipleNodeValues)
-
-  return updateMultipleNodeValues
-}
