--- packages/desktop/tsconfig.types.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.web.json",
-  "compilerOptions": {
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "outDir": "./dist/types",
-    "rootDir": "./src",
-    "skipLibCheck": true
-  },
-  "include": ["src/index.ts", "src/renderer/appStore.ts"],
-  "exclude": ["**/*.test.ts", "**/*.spec.ts", "**/*.stories.ts"]
-}
